+++ app/src/main/java/net/theb2tech/rpg_battle/newHero.java
@@ -88,7 +88,7 @@ public class newHero extends Activity {
 
 
                 hero tempHero = new hero(nameText.getText().toString(),
-                        Double.parseDouble(hpText.getText().toString()),
+                        Integer.parseInt(hpText.getText().toString()),
                         Integer.parseInt(manaText.getText().toString()),
                         Integer.parseInt(wpnDmgText.getText().toString()),
                         Integer.parseInt(defText.getText().toString()),
